+++ src/Concordium/Client/Runner.hs
@@ -907,6 +907,8 @@ processTransactionCmd action baseCfgDir verbose backend =
                     handlePLTUpdateSupply backend baseCfgDir verbose tokenSupplyAction amount tokenId txOpts
                 TransactionPLTModifyList modifyListAction account tokenId txOpts ->
                     handlePLTModifyList backend baseCfgDir verbose modifyListAction account tokenId txOpts
+                TransactionPLTPausation pauseAction tokenId txOpts ->
+                    handlePLTPausation backend baseCfgDir verbose pauseAction tokenId txOpts
 
 -- | Renormalize a 'TokenAmount' to conform to the number of decimal places expected by the
 --  token. If more than the expected number of decimals are given, this fails with an error.
@@ -1063,6 +1065,43 @@ handlePLTModifyList backend baseCfgDir verbose modifyListAction account tokenIdT
         let encodedPayload = Types.encodePayload payload
 
         let nrgCost _ = return $ Just $ tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenListOperationCost
+        txCfg <- liftIO $ getTransactionCfg baseCfg txOpts nrgCost
+
+        let intOpts = toInteractionOpts txOpts
+        let outFile = toOutFile txOpts
+        signAndProcessTransaction_ verbose txCfg encodedPayload intOpts outFile backend
+
+handlePLTPausation ::
+    Backend ->
+    Maybe FilePath ->
+    Bool ->
+    TokenPauseAction ->
+    Text ->
+    TransactionOpts (Maybe Types.Energy) ->
+    IO ()
+handlePLTPausation backend baseCfgDir verbose pauseAction tokenIdText txOpts = do
+    baseCfg <- getBaseConfig baseCfgDir verbose
+    when verbose $ do
+        runPrinter $ printBaseConfig baseCfg
+        putStrLn ""
+
+    withClient backend $ do
+        tokenOperation <- case pauseAction of
+            Pause -> pure CBOR.TokenPause
+            Unpause -> pure CBOR.TokenUnpause
+
+        let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
+        let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
+        let tokenParameter = Types.TokenParameter $ BS.toShort bytes
+
+        tokenId <- case tokenIdFromText tokenIdText of
+            Right val -> return val
+            Left err -> logFatal ["Error couldn't parse token id:", err]
+
+        let payload = Types.TokenUpdate tokenId tokenParameter
+        let encodedPayload = Types.encodePayload payload
+
+        let nrgCost _ = return $ Just $ tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenPauseUnpauseCost
         txCfg <- liftIO $ getTransactionCfg baseCfg txOpts nrgCost
 
         let intOpts = toInteractionOpts txOpts