+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 - The number of decimals in PLT amounts for the `plt` `send`, `mint` and `burn` operations is checked and adjusted if necessary.
   This allows specifying 10 instead of 10.000000 for a PLT with 6 decimal places, for instance.
+- Support PLT pause/unpause operations
 
 ## 9.1.2-alpha (Compatible with node version 9.0.5) - 2025-06-30